requirements.txt
Flask>=3.1.0
python-chess>=1.999
requests>=2.32.0
